ref(reporting): Split conversation reporting by concern
Replace the monolithic conversation reporting module with focused contracts, summary, statistics, transcript, activity, and detail-reader modules. Keep the existing import path as a one-line compatibility facade and preserve the public API.
